Encoder Commit
Changes from the discord version:
- out_enable is held high until in_enable goes high
- latching logic for encoded outputs and output_enable

How this works:
When input goes high, out_enable goes low and the combinational logic is run. On the next posedge, the encoded outputs are latched as well as the out_enable signal. 

-------------------

Testbench was AI generated and just used to show that this version of the encoder works. Use a simulation time around 2000 ns for this testbench to get the outputs for all values of k.